requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
python-dotenv==1.1.0
requests==2.32.3
setuptools==80.4.0
sniffio==1.3.1
starlette==0.46.2
tk==0.1.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
pyarmor==9.1.6
pyarmor.cli.core==7.6.6
